Adding handler for sending Mobiletech (sms) messages
- Added `MobiletechNotificationEvent` for notification
specific to the module - such as response to the
received message.
- Added universal `NotificationHandler` and `SendHandler`
to process outbound messages. All incoming and outgoing
messages are now being stored.
- Added `TestInboundHandler` and `TestNotificationCommand`
to manually test processing of received messages.
See module's README file for more details.
- Added configuration option for definition of free bill key.
This is aimed to be used for generic purpose where we know
that the message is not monetized.
- Fixed inconsistency between modified_at/updated_at columns.

Co-Authored-By: Dominik Skerhak <dominik.skerhak@gmail.com>

remp/crm#1171