Add SME OAuth+Device JWT authorization
Added single JWT-based authorization for SME to comply
with requirement that both OAuth and Device token can
be present within one request and the authorization
should be able to combine them and provide combined
authorization result to the API handler.

Mock implementation is now no longer necessary and is
removed.

remp/crm#1482