Change type of api_logs columns to utf8mb4_unicode_ci
This is to avoid errors caused by funky raw bodies we now log.

The migration is a bit complicated due to blocking-fashion of
the column change. By default we decided to create a new table
and not to migrate the data - it's not really necessary.

If someone decides to migrate with data, they need to enforce
it with the .env variable (see Changelog).

remp/crm#2655