Add LazyApiDecider to prevent unnecessary API handler initialization
The new decider is backwards compatible and if someone registers his
API directly via apiDecider, it will read it from there.

New version also supports HTTP method differentiation for API handlers.
However due to the missing definitions in our modules, we added
a fallback to also try to resolve GET method as a default if we don't
find the match based on the actual HTTP method used for the request.

Co-authored-by: Tibor Mikoczy <mikoczy@efabrica.sk>

remp/crm#2750